Updated intervals to be indepent of "Days" and use Doubles instead of ints.
Created two intervals that can be stored in preferences for notification actions.

+++ FloraSaver/Models/IPlant.cs
@@ -12,15 +12,15 @@
         string GivenName { get; set; }
         int Id { get; set; }
         string ImageLocation { get; set; }
-        int? MistInterval { get; set; }
+        double? MistInterval { get; set; }
         string PlantSpecies { get; set; }
-        int? SunInterval { get; set; }
+        double? SunInterval { get; set; }
         TimeSpan TimeOfLastMisting { get; set; }
         TimeSpan TimeOfLastMove { get; set; }
         TimeSpan TimeOfLastWatering { get; set; }
         TimeSpan TimeOfNextMisting { get; set; }
         TimeSpan TimeOfNextMove { get; set; }
         TimeSpan TimeOfNextWatering { get; set; }
-        public int? WaterInterval { get; set; }
+        public double? WaterInterval { get; set; }
     }
 }
